Refactor VCR implementation to comply with spec
- Extract VcrReplayMismatchError to dedicated errors.py module
- Replace ValueError with VcrReplayMismatchError across cassette, filesystem, and mcp_server modules
- Rename FilesystemAccess protocol to FilesystemAccessProtocol for clarity
- Remove backwards-compat "index" key from AutoStepEvent.from_dict
- Remove backwards-compat finish() alias from VcrMekaraServer
- Move MekaraServerTestDriver from tests/vcr_test_driver.py into src/mekara/vcr/mcp_server.py
- Delete vcr_test_driver.py module
